#962 Move giving cards to user registration
